added the window abstraction into its own window class
