+++ src/main.cpp
@@ -17,8 +17,6 @@ int main() {
     }
 
 
-
-
     float vertices[] = {
         // positions        // colors
          0.0f,  0.5f, 0.0f,  1.0f, 0.0f, 0.0f,
